build module done & merge user client for picket(aslan)
