--- pkg/microservice/picket/client/user/user.go
@@ -1,58 +1,0 @@
-/*
-Copyright 2021 The KodeRover Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package user
-
-import (
-	"net/http"
-	"net/url"
-
-	"github.com/koderover/zadig/pkg/tool/httpclient"
-	"github.com/koderover/zadig/pkg/types"
-)
-
-func (c *Client) DeleteUser(userId string, header http.Header, qs url.Values) ([]byte, error) {
-	url := "/users/" + userId
-
-	res, err := c.Delete(url, httpclient.SetHeadersFromHTTPHeader(header), httpclient.SetQueryParamsFromValues(qs))
-	if err != nil {
-		return nil, err
-	}
-
-	return res.Body(), nil
-}
-
-func (c *Client) SearchUsers(header http.Header, qs url.Values, body interface{}) (*types.UsersResp, error) {
-	url := "/users/search"
-	result := &types.UsersResp{}
-	_, err := c.Post(url, httpclient.SetHeadersFromHTTPHeader(header), httpclient.SetQueryParamsFromValues(qs), httpclient.SetBody(body), httpclient.SetResult(result))
-	if err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-type SearchArgs struct {
-	Name         string   `json:"name,omitempty"`
-	Account      string   `json:"account,omitempty"`
-	IdentityType string   `json:"identity_type,omitempty"`
-	UIDs         []string `json:"uids,omitempty"`
-	PerPage      int      `json:"per_page,omitempty"`
-	Page         int      `json:"page,omitempty"`
-}
-
-type SearchRes struct {
-}
